pyproject.toml
[project]
name = "maiar-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mypy>=1.16.1",
    "ruff>=0.12.3",
    "gitingest>=0.1.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.8",
    "langchain-text-splitters>=0.0.1",
    "chromadb>=0.4.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]
[tool.ruff]
# Target Python 3.10+
target-version = "py310"

# Set line length
line-length = 88

# Exclude files/directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "migrations/",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

# Ignore specific rules if needed
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
# Basic settings
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

# Strictness settings
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

# Error settings
show_error_codes = true
show_column_numbers = true
pretty = true

# Import settings
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "gitingest"
ignore_missing_imports = true

# Import settings
follow_imports = "normal"
